requirements.txt
Flask
gTTS
twilio
requests
beautifulsoup4
schedule
pyngrok
python-dotenv